Add validate module to encapsulate validations
