+++ lib/cli/input/init.js
@@ -1,25 +1,18 @@
 const inquirer = require('inquirer');
 const PromptGenerator = require('../../utils/PromptGenerator');
-
-const getLengthValidator = invalidMessage => (value) => {
-  if (value.length) {
-    return true;
-  }
-
-  return invalidMessage;
-};
+const initValidator = require('../../validate/init');
 
 const usernamePromptGenerator = new PromptGenerator();
 usernamePromptGenerator.addProperty('name', 'username');
 usernamePromptGenerator.addProperty('type', 'input');
 usernamePromptGenerator.addProperty('message', 'Enter your AutolabJS username:');
-usernamePromptGenerator.addProperty('validate', getLengthValidator('Please enter your username'));
+usernamePromptGenerator.addProperty('validate', initValidator.getLengthValidator('Please enter your username'));
 
 const passwordPromptGenerator = new PromptGenerator();
 passwordPromptGenerator.addProperty('name', 'password');
 passwordPromptGenerator.addProperty('type', 'password');
 passwordPromptGenerator.addProperty('message', 'Enter your AutolabJS password:');
-passwordPromptGenerator.addProperty('validate', getLengthValidator('Please enter your password'));
+passwordPromptGenerator.addProperty('validate', initValidator.getLengthValidator('Please enter your password'));
 
 const credentials = [
   usernamePromptGenerator.getPrompt(),
