+++ lib/cli/input/prefs.js
@@ -7,6 +7,7 @@ const defaultPrefs = JSON.parse(require('fs').readFileSync(defaultPrefPath, 'utf
 const { supportedLanguages } = defaultPrefs;
 const { logger } = require('../../utils/logger');
 const prefPrompts = require('./prefsprompt');
+const prefsValidator = require('../../validate/prefs');
 
 const changeLangFlag = (lang) => {
   if (supportedLanguages.indexOf(lang) === -1) {
@@ -34,13 +35,13 @@ const changeLang = async (options) => {
 };
 
 const changeMainServerFlag = (host, port) => {
-  if (!prefPrompts.hostValidator(host, true)) {
+  if (!prefsValidator.hostValidator(host, true)) {
     logger.moduleLog('debug', 'Prefs Input', `Invalid mainserver host ${host}`);
     return {
       name: 'invalid_host',
     };
   }
-  if (!prefPrompts.portValidator(port, true)) {
+  if (!prefsValidator.portValidator(port, true)) {
     logger.moduleLog('debug', 'Prefs Input', `Invalid mainserver port ${host}`);
     return {
       name: 'invalid_port',
@@ -57,7 +58,7 @@ const changeMainServerFlag = (host, port) => {
 };
 
 const changeGitlabFlag = (host) => {
-  if (!prefPrompts.hostValidator(host, true)) {
+  if (!prefsValidator.hostValidator(host, true)) {
     logger.moduleLog('debug', 'Prefs Input', `Invalid Gitlab host ${host}`);
     return {
       name: 'invalid_host',
@@ -113,7 +114,7 @@ const changeLogger = async (options) => {
     return prefPrompts.changeLoggerPrompt();
   }
 
-  if (blacklist && !prefPrompts.keywordValidator(blacklist, true)) {
+  if (blacklist && !prefsValidator.keywordValidator(blacklist, true)) {
     return {
       name: 'invalid_blacklist_keyword',
     };
