+++ lib/cli/input/prefsprompt.js
@@ -2,9 +2,8 @@ const inquirer = require('inquirer');
 const path = require('path');
 const fs = require('fs');
 
-const validator = require('validator');
 const PromptGenerator = require('../../utils/PromptGenerator');
-const preferenceManager = require('../../utils/preference-manager');
+const prefsValidator = require('../../validate/prefs');
 
 const defaultPrefPath = path.join(__dirname, '../../../default-prefs.json');
 const defaultPrefs = JSON.parse(fs.readFileSync(defaultPrefPath, 'utf8'));
@@ -27,28 +26,6 @@ const changeLangPrompt = async () => {
   };
 };
 
-const hostValidator = (host, flags) => {
-  if (validator.isFQDN(host) || validator.isIP(host)) {
-    return true;
-  }
-  return flags === true ? false : 'Enter a valid hostname or IP';
-};
-
-const portValidator = (port, flags) => {
-  if (validator.isPort(port)) {
-    return true;
-  }
-  return flags === true ? false : 'Enter a valid port';
-};
-
-const keywordValidator = (keyword, flags) => {
-  const { blacklist } = preferenceManager.getPreference({ name: 'cliPrefs' }).logger;
-  if (!blacklist.includes(keyword)) {
-    return true;
-  }
-  return flags === true ? false : 'Keyword already exists. Enter any other keyword';
-};
-
 const getPromptGenerator = (name, type, message, validate) => {
   const promptGenerator = new PromptGenerator();
   promptGenerator.addProperty('name', name);
@@ -65,12 +42,12 @@ const getTypePromptGenerator = (name, type, message, validate, choices) => {
 };
 
 const getHostPrompt = () => {
-  const hostPromptGenerator = getPromptGenerator('host', 'input', 'Enter the host:', hostValidator);
+  const hostPromptGenerator = getPromptGenerator('host', 'input', 'Enter the host:', prefsValidator.hostValidator);
   return hostPromptGenerator.getPrompt();
 };
 
 const getPortPrompt = () => {
-  const portPromptGenerator = getPromptGenerator('port', 'input', 'Enter the port:', portValidator);
+  const portPromptGenerator = getPromptGenerator('port', 'input', 'Enter the port:', prefsValidator.portValidator);
   return portPromptGenerator.getPrompt();
 };
 
@@ -133,7 +110,8 @@ const changeMaxSize = async () => {
 };
 
 const changeBlacklist = async () => {
-  const blacklistPromptGenrator = getPromptGenerator('keyword', 'input', 'Enter the keyword to be added to logger blacklist:', keywordValidator);
+  const blacklistPromptGenrator = getPromptGenerator('keyword', 'input', 'Enter the keyword to be added to logger blacklist:',
+    prefsValidator.keywordValidator);
   const blacklistPrefs = await inquirer.prompt(blacklistPromptGenrator.getPrompt());
 
   return {
@@ -159,9 +137,6 @@ const changeLoggerPrompt = async () => {
 
 module.exports = {
   changeLangPrompt,
-  hostValidator,
-  portValidator,
-  keywordValidator,
   changeMainServerPrompt,
   changeGitlabPrompt,
   getServerTypePrompt,
