+++ lib/controller/prefs.js
@@ -2,20 +2,14 @@ const prefsInput = require('../cli/input/prefs');
 const prefsOutput = require('../cli/output/prefs');
 const prefsModel = require('../model/prefs');
 
-const validator = require('./validation');
-
 const onPrefs = async (args, options, logger) => {
   logger.moduleLog('info', 'Prefs', 'Prefs command invoked for');
   logger.moduleLog('info', 'Prefs', args.preference);
-  if (validator.prefs(args, options)) {
-    let changePrefEvent = await prefsInput.getInput(args, options);
-    logger.moduleLog('debug', 'Prefs', 'Event for the prefs command called');
-    logger.moduleLog('debug', 'Prefs', changePrefEvent);
-    changePrefEvent = await prefsModel.storePrefs(changePrefEvent);
-    prefsOutput.sendOutput(changePrefEvent);
-  } else {
-    logger.moduleLog('error', 'Prefs', 'Invalid command');
-  }
+  let changePrefEvent = await prefsInput.getInput(args, options);
+  logger.moduleLog('debug', 'Prefs', 'Event for the prefs command called');
+  logger.moduleLog('debug', 'Prefs', changePrefEvent);
+  changePrefEvent = await prefsModel.storePrefs(changePrefEvent);
+  prefsOutput.sendOutput(changePrefEvent);
 };
 
 const addTo = (program) => {
