+++ lib/validate/init.js
@@ -1,0 +1,11 @@
+const getLengthValidator = invalidMessage => (value) => {
+  if (value.length) {
+    return true;
+  }
+
+  return invalidMessage;
+};
+
+module.exports = {
+  getLengthValidator,
+};
